requirements.txt
pdfplumber
llama-cpp-python==0.3.16
